Strange sky effect at 90 radians?  Nope!
The Skybox now handles the sky itself now, though - sadly I can't fully
resolve the circle lighting issue when the player looks ~90 radians.
This issue exists because of the illusion fog effects and lighting
differences which does not easily blend well for proper consistency.
However, I applied a nasty hack in attempt to 'cover' atleast 90% of its
existence.  It should not be an eyesore any longer and is only
noticeable when the user is deliberate looking for it specifically.  The
hack is marked and tagged using GZDB's UDMF UI if anyone else has any
better ideas.