Revised map script #Includes
Instead of using an absolute path to point to a required file, it'll use
a relative path - assuming that the map's assets in GZDoom Builder has
the directory named 'Scripts' available in the maps resources.  This
will work on abroad filesystems while using GZDoom Builder and not just
only my individual desktop system.