Generalized revision
Expunged Subversion code and replenished with Git support and
functionality.
Renamed the Default directory to 'TGRDM3' instead of 'Morgenstern'.
Revised the change log hard limit variable name to be generally vague
but yet exactly states its purpose.  The value is still the same.
Removed SVNRevisionOld, SVNRevisionNew, and SVNRevisionRange.
Project (not program) version is now: 1
ProjectNameCompact is now 'TGRDM3'
Updated the ProjectVersion and ReleaseDate.