GitHub Services directory
Since we can not explicitly upload images to the Wiki, this directory
will hold as a container instead.  This will allow us to use the
required data as we need it on the wiki.

Note: I could very well and upload the images to another medium outside
of the project data, however, I can't be certain that years later - that
same target will still exist.  This seems to be the better route, as the
project will still be around at that giving time.