Added SkullTag's F1 Help Screens [DM, Duel]
Included SkullTag's F1 Help Screens for Deathmatch and Duel.  This works
for both (G)ZDoom and SkullTag's derivatives.  The documentation is
severely restricted as the Game Mode that is set in SkullTag (or its
derivatives) will merely be enforced to use the pre-set F1 definition
per map and not plausible to inter-change automatically.