Improved text2 standard and readability
It is now possible to leave markers like: TODO, FIX, and LOOK, the text
will appear red with an underline.  This will only take place for
comments.
Support for folding brackets is now possible.
When using #if, #elseif is now supported for middle seting, to close -
still use #endif.
Numbers containing with or without decimals will have a different colour
property.