Enforce Sprite view angles
Enforce [Y|XY] billboard settings.  This is useful for viewing sprites
at extreme angles.  Necessary sprites use XY viewing angles, while some
that are not as important are only viewable at the Y-axis.  In addition,
this is necessary as the end-users can be able to disable voxels - which
sprites are the secondary option after Voxels.