Globalized necessary variables
MusicPlayList array is now globalized.
Added two new global variables: MusicPlayListReady [int] is useful only
for when determining weither the state of the play list array is ready
for use or needs to be re-generated.  In addition
'MusicPlayListHighlight' stores the current position of the array index
that is being played.