Generalized Revision [WIP]
Renamed function 'ReorganizeArray' -> 'MusicPlayer_ShufflePlayList'
Added a new function 'GenerateMusicPlaylist', which makes sure that the
array is ready for use.
MusicPlayer is now tagged as 'OPEN', this now executes immediately at
start.
Revised MusicPlayer script to work with the current state of this entire
script file.  This is merely a front-end access point to launch this
feature at each map start.