Added an ACS script to detect game mode
I have added a test ACS script to try to detect the game mode; however,
this is not a perfect solution - but this is the only way it'll work
with GZDoom && Zandronum with minimal issues.  Also, the script uses an
inversion to GAME_NET_TEAMGAME.  Meaning, the script - to work with
GZDoom, the script will check if GAME_SINGLE_PLAYER,
GAME_NET_COOPERATIVE, GAME_NET_DEATHMATCH, and GAME_TITLE_MAP are all
false - and assumes we're in Domination in this case.  The same effect
would happen if we used GAME_NET_TEAMGAME, just nicer in the code.  To
point out the issues, if the the gameserver or client uses CTF or any
other team game mode, this script will think we're in Domination.  This
is a limitation and probably not fixable without further implementations
to the engines.