Revised GZDB documentation and added delay
I revised the accuracy of the GZDB documentation.
I added a new feature to this code, by adding the first argument which
issues a custom 'delay'.  The default delay is 4, which matches the
original code.  However, it is now possible to slow it down or speed up
the delay.  This feature allows the map developer to fully customize the
delay in anyway they see fit - provided it is positive number.