Redoing the map pictures again
Since (G)ZDoom can use Graphics as Textures, all of TX6_* has been
dropped for the Intermission graphics.  The images are still under JPG
and at the same 90% Quality, but with smoothing - to hide the artifacts

Relates to this ticket:
https://github.com/SibTiger/TGRDM3/issues/35
a bit more, but other parts of the images will also be smoothed out
aswell -- this technique fits as a blur to the background, and offers
attention to the foreground text.