Railgun will no longer check for reload
Change of heart with this implementation.  I particularly like the
reload mechanism as it does not have that proper 'flow' as opposed to
the other weapons used in this project.  However, if others like the
reload checking - the instructions still exist and can easily be changed
by uncommenting the check instruction and comment out the 'Goto
ReadyToUse' at line 85.